refector : type 정리, providers 분리, 게시글 버그 수정, axios instance 호출  합치기, 일부 css 수정
