+++ src/apis/user.ts
@@ -1,8 +1,6 @@
-import { getMyInfoType } from 'types'
 import { getInstance } from './client'
 
-export const getMyInfoAPI: getMyInfoType = () => {
-  return getInstance(true)
-    .get('/myInfo')
-    .then((res) => res.data)
+export const getMyInfoAPI = async () => {
+  const res = await getInstance(true).get('/myInfo')
+  return res.data
 }
