+++ src/components/Card/card.module.scss
@@ -2,6 +2,7 @@
 @use 'src/styles/mixins/vwConvert';
 @use 'src/styles/constants/sizes';
 @use 'src/styles/constants/colors';
+@use '/src/styles/mixins/typography';
 
 .card {
   position: relative;
@@ -59,6 +60,12 @@
     width: 104px;
     height: 95px;
   }
+
+  .market {
+    @include typography.lineClamp(1, 18);
+    width: 70%;
+  }
+
   @media (max-width: sizes.$SCREEN_SM) {
     @include vwConvert.mo(padding, 12 21);
     @include vwConvert.mo(border-radius, 15);
