+++ src/components/Slider/CategorySlider.tsx
@@ -1,4 +1,4 @@
-import React, { Dispatch, SetStateAction } from 'react'
+import { Dispatch, SetStateAction } from 'react'
 import { category } from 'types'
 import SliderWrapper from './SliderWrapper'
