+++ src/pages/Posting/TimePicker/index.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction, useEffect } from 'react'
 import cx from 'classnames'
 import { useRecoilValue, useRecoilState } from 'recoil'
-import { postingDateState, postingTimeState, modalChangeState, isEditDefaultValue } from 'recoil/post.atom'
+import { postingDateState, postingTimeState, postingModalState, isEditDefaultValue } from 'recoil/post.atom'
 
 import styles from './timePicker.module.scss'
 
@@ -14,7 +14,7 @@ const TimePicker = ({ onClick, propTime }: Props) => {
   const isDefaultValue = useRecoilValue(isEditDefaultValue)
   const [time, setTime] = useRecoilState(postingTimeState)
   const [date, setDate] = useRecoilState(postingDateState)
-  const valueUpdate = useRecoilValue(modalChangeState)
+  const valueUpdate = useRecoilValue(postingModalState)
 
   // 상세에서 불러온 데이터 일 때만 실행
   useEffect(() => {
@@ -36,7 +36,7 @@ const TimePicker = ({ onClick, propTime }: Props) => {
     <div className={styles.timePicker}>
       <button
         type='button'
-        className={cx(styles.popupBtn, { [styles.update]: valueUpdate })}
+        className={cx(styles.popupBtn, { [styles.update]: valueUpdate || propTime })}
         onClick={() => onClick(true)}
       >
         <span className={styles.label}>만날 시간</span>
