+++ src/providers/query.tsx
@@ -1,0 +1,22 @@
+import { ReactNode } from 'react'
+import { QueryClient, QueryClientProvider } from 'react-query'
+import { ReactQueryDevtools } from 'react-query/devtools'
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      refetchOnMount: false,
+      refetchOnWindowFocus: false,
+      retry: 1,
+      // eslint-disable-next-line no-console
+      onError: (error) => console.log((error as any).message),
+    },
+  },
+})
+
+export const QueryProvider = ({ children }: { children: ReactNode }) => (
+  <QueryClientProvider client={queryClient}>
+    <ReactQueryDevtools initialIsOpen={false} />
+    {children}
+  </QueryClientProvider>
+)
