+++ src/styles/mixins/_position.scss
@@ -17,7 +17,7 @@
   @return type-of($value) != 'null' and
     (
       str-slice($value + '', 1, 4) == 'calc' or index(auto inherit initial 0, $value) or
-        (type-of($value) == 'number' and not(unitless($value)))
+      (type-of($value) == 'number' and not(unitless($value)))
     );
 }
