+++ src/types/modal.ts
@@ -1,0 +1,6 @@
+/* modal show, close type */
+export interface ModalPropsType {
+  show: boolean
+  close: () => void
+  isEdit?: boolean
+}
