+++ src/types/user.ts
@@ -1,0 +1,15 @@
+export interface myInfo {
+  nickname: string
+  email: string
+}
+
+/* my-info type */
+export interface myInfoType {
+  userId: number
+  nickname: string
+  email: string
+  location: string
+  profileUrl: string
+  lat: string
+  lon: string
+}
